Crawler.java and App.java
App.java
Changed so crawl() is now crawl(params)

Crawler.java
parses html to xml.

Things need to work on:
1) depth. Need to now search deeper but will work on soon
2) printing nicer format of xml

How to use crawler:
1) need to have file.xml on desktop or somewhere and change path to
where you want xml to be
2) needs to have --depth (number) for example --depth1
3) i commented out a method called printXml() so you can uncomment to
see it be printed in console

Any other suggestions or comments? Lemme know if it works for you at
least and will continue when it works for u guys